small bug fixed

+++ Network Tomography Toolbox/Codes/Inference_Tomography.py
@@ -146,8 +146,8 @@ for n_ind in neuron_range:
         W_inferred = np.reshape(W_inferred,[no_neurons,1])
     
     #.........................Save the Belief Matrices.........................
-    file_name_ending = 'I_' + str(inference_method) + '_S_' + str(sparsity_flag*sparse_thr0) + '_T_' + str(T)
-    file_name_ending = file_name_ending + '_C_' + str(num_process) + '_B_' + str(block_size)
+    file_name_ending = 'I_' + str(inference_method) + '_S_' + str(float(sparsity_flag*sparse_thr0)) + '_T_' + str(int(T))
+    file_name_ending = file_name_ending + '_C_' + str(int(num_process)) + '_B_' + str(int(block_size))
     file_name_ending = file_name_ending + '_K_' + kernel_choice + '_H_' + str(class_sample_freq)
     #file_name_ending = file_name_ending + '_F_' + str(class_sample_freq)
